modify CRUD of some tables

+++ src/models/Account.js
@@ -3,6 +3,8 @@ module.exports = (sequelize, DataTypes) => {
     class Account extends Model {
         static associate(models) {
           // define association here
+          Account.hasOne(models.User, {foreignKey: 'user_id'});
+          Account.hasOne(models.Admin, {foreignKey: 'admin_id'});
         }
         // CREATE operation, role: 'admin','user','staff'
         static async createAccount(c_username, c_password, c_role, c_email) {
@@ -18,6 +20,18 @@ module.exports = (sequelize, DataTypes) => {
                 email: c_email
             });
 
+            if (c_role == 'user')
+            {
+              // insert into User table
+              const newUser = await this.associations.User.target.createUser({
+                user_id : newAccount.dataValues.account_id
+              });
+            }
+            else if (c_role == 'admin')
+            {
+              // insert into Admin table
+              const newAdmin = await this.associations.Admin.target.createAdmin(newAccount.dataValues.account_id, '');
+            }
             return newAccount;
         }
 
@@ -71,10 +85,24 @@ module.exports = (sequelize, DataTypes) => {
 
         // DELETE operation
         static async deleteAccount(accountId) {
-            const deletedRowCount = await Account.destroy({
-                where: { account_id: accountId }
+            let deletedAcc = await Account.getAccountById(accountId);
+            if (deletedAcc == null) 
+            {
+              console.log("No account with this ID is found");
+              return;
+            }
+            if (deletedAcc.dataValues.role == 'user')
+            {
+              const deleteUser = await this.associations.User.target.deleteUser(deletedAcc.dataValues.account_id);
+            }
+            else if (deletedAcc.dataValues.role == 'admin')
+            {
+              const deleteUser = await this.associations.Admin.target.deleteAdmin(deletedAcc.dataValues.account_id);
+            }
+            deletedAcc = await Account.destroy({
+              where: { account_id: accountId }
             });
-            return deletedRowCount > 0;
+            return deletedAcc > 0;
         }
       }
       Account.init({
