+++ src/models/User.js
@@ -1,33 +1,19 @@
 const { Model } = require('sequelize');
-// const SubscriptionPlan = require('./SubscriptionPlan');
-
 module.exports = (sequelize, DataTypes) => {
     class User extends Model {
         static associate(models) {
             User.belongsTo(models.Account, { foreignKey: 'user_id' });
-            User.hasOne(models.SubscriptionPlan, {foreignKey: 'plan_id'});
+            User.hasOne(models.SubscriptionPlan, {foreignKey: 'user_id'});
         }
-        static async createUser(c_account ,c_name, c_birthday, c_parental_mode) {
-            // const defaultPlan = await SubscriptionPlan.createSubscriptionPlan(null, null, 1);
-            const defaultPlan = await this.associations.SubscriptionPlan.target.createSubscriptionPlan(null, null, 1);
-            // Call the stored procedure
-            // const result = await sequelize.query('CALL createDefaultSubscriptionPlan()', {
-            //     replacements: [c_name, c_age, c_birthday, c_parental_mode],
-            //     type: sequelize.QueryTypes.CALL
-            // });
-            // const plan = result[0][0];
-
-            const now = new Date();
-            const dateObject = new Date(c_birthday);
-            const c_age = now.getFullYear() - dateObject.getFullYear();
-            const newUser = await User.create({
-                user_id: c_account,
-                name: c_name,
-                age: c_age,
-                birthday: c_birthday,
-                parental_mode: c_parental_mode,
-                plan_id: defaultPlan.dataValues.plan_id
-            });
+        static async createUser(UserData) {
+            const newUser = await User.create(UserData);
+            // first user account to be created will be create with Free Subscription Plan, they can upgrade later
+            const defaultPlan = await this.associations.SubscriptionPlan.target.createSubscriptionPlan(
+                {
+                    user_id: newUser.dataValues.user_id,
+                    subscription_id: 1
+                }
+            );
             return newUser;
         }
 
@@ -61,11 +47,18 @@ module.exports = (sequelize, DataTypes) => {
         }
 
         static async deleteUser(userId) {
+            let deletedUser = await User.getUserById(userId);
+            if (deletedUser == null) 
+            {
+                console.log("No user found");
+                return;
+            }
+            const plan = await User.associations.SubscriptionPlan.target.deleteSubscriptionPlan(userId);
             try {
-                const deletedRowCount = await User.destroy({
+                deletedUser = await User.destroy({
                     where: { user_id: userId }
                 });
-                return deletedRowCount > 0;
+                return deletedUser > 0;
             } catch (error) {
                 throw new Error(`Error deleting user: ${error.message}`);
             }
@@ -81,12 +74,22 @@ module.exports = (sequelize, DataTypes) => {
             type: DataTypes.STRING(255),
             charset: 'utf8mb4'
         },
-        age: DataTypes.INTEGER,
-        birthday: DataTypes.DATE,
-        parental_mode: DataTypes.INTEGER,
-        plan_id: {
+        age: 
+        {
             type: DataTypes.INTEGER,
-            allowNull: false
+            get() 
+            {
+                const now = new Date();
+                const birthDate = new Date(this.getDataValue('birthday'));
+                const c_age = now.getFullYear() - birthDate.getFullYear();
+                return c_age;
+            }
+        },
+        birthday: DataTypes.DATE,
+        parental_mode: 
+        {
+            type: DataTypes.INTEGER,
+            defaultValue: 0
         }
     }, {
         sequelize,
